+++ test/templates/website/design.js
@@ -1,4 +1,4 @@
-const {cx, Locale, bsiProperty} = require('@bsi-cx/design-build');
+const {cx, Locale, bsiProperty, WebsiteContentType} = require('@bsi-cx/design-build');
 
 /**
  * @type {string}
@@ -58,11 +58,16 @@ module.exports = cx.design
           .withName('Vorlage für Inhaltsseiten')
           .withEditable(true)
           .withFile(require('./includes/page.html')),
+          cx.pageInclude
+            .withName('Vorlage für Blog')
+            .withEditable(true)
+            .withFile(require('./includes/page.html'))
+            .withWebsiteContentType(WebsiteContentType.CUSTOMER_PARTNER),
         cx.include
           .withIdentifier('pagination-element')
           .withName('Pagination')
           .withEditable(false)
-          .withContentType('pre-defined')
+          .withIncludeType('pre-defined')
           .withFile(require('./includes/pagination-element.hbs')),
         cx.include
           .withIdentifier('footer')
@@ -79,4 +84,5 @@ module.exports = cx.design
           .withName('Navigation')
           .withEditable(false)
           .withFile(require('./includes/navigation.hbs'))))
-  .withNLS(Object.values(require('./nls')));
+  .withNLS(Object.values(require('./nls')))
+  .withWebsiteContentTypes(WebsiteContentType.BLOG, WebsiteContentType.CUSTOMER_PARTNER);